python-examples/requirements.txt
# Python dependencies for HWPX to Markdown conversion scripts
JPype1>=1.4.1
markdown>=3.4.0